+++ src/helper.ts
@@ -1,5 +1,4 @@
-import { firestore } from 'firebase-admin'
-
+import { GeoPoint } from "firebase-admin/firestore"
 
 export interface IImportOptions {
   dates?: string[]
@@ -14,8 +13,9 @@ export interface IExportOptions {
   docsFromEachCollection?: number
   refs?: string[]
   queryCollection?: (
-    ref: firestore.CollectionReference<firestore.DocumentData>
-  ) => Promise<firestore.QuerySnapshot<any>>
+    ref: FirebaseFirestore.CollectionReference
+  ) => Promise<FirebaseFirestore.QuerySnapshot>
+  
 }
 
 export const makeGeoPoint = (geoValues: {
@@ -26,7 +26,7 @@ export const makeGeoPoint = (geoValues: {
     return null
   }
 
-  return new firestore.GeoPoint(geoValues._latitude, geoValues._longitude)
+  return new GeoPoint(geoValues._latitude, geoValues._longitude)
 }
 
 /**